LOGICA INICIAL CREADA Y OPTIMIZADA v2
